+++ bireme/leisref/forms.py
@@ -11,7 +11,7 @@ from django.utils.translation import ugettext_lazy as _
 from main.models import Descriptor, Keyword, ResourceThematic
 from attachments.models import Attachment
 
-from utils.forms import DescriptorRequired, ResourceThematicRequired
+from utils.forms import BaseDescriptorInlineFormSet, ResourceThematicRequired
 from models import *
 
 import simplejson
@@ -133,8 +133,14 @@ class ActRelatedForm(forms.ModelForm):
 
 
 # definition of inline formsets
-DescriptorFormSet = generic_inlineformset_factory(Descriptor, form=DescriptorForm,
-                                                  exclude=('status',), can_delete=True, extra=1)
+DescriptorFormSet = generic_inlineformset_factory(
+    Descriptor,
+    form=DescriptorForm,
+    formset=BaseDescriptorInlineFormSet,
+    exclude=('status',),
+    can_delete=True,
+    extra=1
+)
 
 AttachmentFormSet = generic_inlineformset_factory(Attachment, form=AttachmentForm,
                                                   exclude=('short_url',), can_delete=True, extra=1)
