+++ bireme/utils/tests.py
@@ -1,9 +1,13 @@
 # coding: utf-8
+from django.contrib.auth.models import User
+from django.contrib.contenttypes.forms import generic_inlineformset_factory
 from django.test import TestCase
 from django.test.client import Client
 from django.test.utils import override_settings
 
-from django.contrib.auth.models import User
+from main.models import Descriptor
+from utils.forms import BaseDescriptorInlineFormSet
+
 
 @override_settings(AUTHENTICATION_BACKENDS=('django.contrib.auth.backends.ModelBackend',))
 class BaseTestCase(TestCase):
@@ -80,3 +84,56 @@ class BaseTestCase(TestCase):
         # only superuser can edit lists
         user_admin = User.objects.create_superuser('admin', 'admin@test.com', 'admin')
         self.client.login(username='admin', password='admin')
+
+
+class DescriptorFormSetTest(BaseTestCase):
+    def setUp(self):
+        self.DescriptorFormSet = generic_inlineformset_factory(
+            Descriptor,
+            formset=BaseDescriptorInlineFormSet,
+            exclude=('status',),
+            can_delete=True,
+            extra=1
+        )
+
+    def test_duplicated(self):
+        """Duplicated descriptors texts should not be accepted"""
+        data = {
+            'main-descriptor-content_type-object_id-TOTAL_FORMS' : '2',
+            'main-descriptor-content_type-object_id-INITIAL_FORMS': '0',
+            'main-descriptor-content_type-object_id-MAX_NUM_FORMS': '',
+
+            'main-descriptor-content_type-object_id-0-id' : '',
+            'main-descriptor-content_type-object_id-0-text' : 'malaria',
+            'main-descriptor-content_type-object_id-0-code' : '^d8462',
+            'main-descriptor-content_type-object_id-0-status' : '0',
+
+            'main-descriptor-content_type-object_id-1-id' : '',
+            'main-descriptor-content_type-object_id-1-text' : 'malaria',
+            'main-descriptor-content_type-object_id-1-code' : '^d8462',
+            'main-descriptor-content_type-object_id-1-status' : '0',
+        }
+
+        formset = self.DescriptorFormSet(data)
+        self.assertFalse(formset.is_valid())
+
+    def test_unique_text(self):
+        """Unique descriptors texts should validate form"""
+        data = {
+            'main-descriptor-content_type-object_id-TOTAL_FORMS' : '2',
+            'main-descriptor-content_type-object_id-INITIAL_FORMS': '0',
+            'main-descriptor-content_type-object_id-MAX_NUM_FORMS': '',
+
+            'main-descriptor-content_type-object_id-0-id' : '',
+            'main-descriptor-content_type-object_id-0-text' : 'malaria',
+            'main-descriptor-content_type-object_id-0-code' : '^d8462',
+            'main-descriptor-content_type-object_id-0-status' : '0',
+
+            'main-descriptor-content_type-object_id-1-id' : '',
+            'main-descriptor-content_type-object_id-1-text' : 'ANATOMY',
+            'main-descriptor-content_type-object_id-1-code' : '^d59005',
+            'main-descriptor-content_type-object_id-1-status' : '0',
+        }
+
+        formset = self.DescriptorFormSet(data)
+        self.assertTrue(formset.is_valid())
